+++ Example/AuthExample/HomeViewController.swift
@@ -7,109 +7,35 @@
 //
 
 import UIKit
+import WebKit
 import ConsentViewController
 
-class HomeViewController: UIViewController, GDPRConsentDelegate {
-    lazy var consentViewController: GDPRConsentViewController = { return GDPRConsentViewController(
-        accountId: 22,
-        propertyId: 7639,
-        propertyName: try! GDPRPropertyName("tcfv2.mobile.webview"),
-        PMId: "122058",
-        campaignEnv: .Public,
-        consentDelegate: self
-    )}()
+class HomeViewController: UIViewController {
+    static let webviewUrl = URL(string: "http://localhost:8080/")!
+    static let notFoundHtml = Bundle.main.path(forResource: "webserver/404", ofType: "html")!
 
-    var authId = ""
+    @IBOutlet weak var webview: WKWebView!
 
-    @IBOutlet var authIdLabel: UILabel!
-    @IBOutlet var consentTableView: UITableView!
-
-    let tableSections = ["userData", "consents"]
-    var userData: [String] = []
-    var consents: [String] = []
-
-    func gdprConsentUIWillShow() {
-        self.present(consentViewController, animated: true, completion: nil)
-    }
-
-    func consentUIDidDisappear() {
-        self.dismiss(animated: true, completion: nil)
-    }
-
-    func onConsentReady(gdprUUID: GDPRUUID, userConsent: GDPRUserConsent) {
-        self.userData = [
-            "gdprUUID: \(gdprUUID)",
-            "consent string: \(userConsent.euconsent)"
-        ]
-        self.consents =
-            userConsent.acceptedVendors.map({ v in return "Vendor: \(v)"}) +
-            userConsent.acceptedCategories.map({ c in return "Purpose: \(c)"})
-        self.consentTableView.reloadData()
-    }
-
-    func onError(error: GDPRConsentViewControllerError?) {
-        print(error.debugDescription)
-    }
+    var authId: String!
 
     override func viewDidLoad() {
-        super.viewDidLoad()
-        authIdLabel.text = authId
-        initData()
-        consentViewController.loadMessage(forAuthId: authId)
-    }
-
-    @IBAction func onSettingsPress(_ sender: Any) {
-        initData()
-        consentViewController.loadPrivacyManager()
+        webview.setConsentFor(authId: authId) { error in print(error) }
+        webview.navigationDelegate = self
+        webview.load(URLRequest(url: HomeViewController.webviewUrl))
     }
 }
 
-extension HomeViewController: UITableViewDataSource {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        switch section {
-        case 0:
-            return userData.count
-        case 1:
-            return consents.count
-        default:
-            return 0
+extension HomeViewController: WKNavigationDelegate {
+    func webView(_ webView: WKWebView, didFailProvisionalNavigation navigation: WKNavigation!, withError error: Error) {
+        webview.loadHTMLString(try! String(contentsOfFile: HomeViewController.notFoundHtml), baseURL: nil)
+    }
+
+    /// Get the value of authId cookie from the document
+    func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
+        webView.getAuthId { authId, error in
+            error != nil ?
+                print(error.debugDescription) :
+                print("AuthId:", authId ?? "No AuthId")
         }
-    }
-
-    func initData() {
-        self.userData = [
-            "gdprUUID: loading...",
-            "consent string: loading..."
-        ]
-        consentTableView.reloadData()
-    }
-
-    func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
-        return tableSections[section]
-    }
-
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return tableSections.count
-    }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "PlainCell", for: indexPath)
-        switch indexPath.section {
-        case 0:
-            cell.textLabel?.adjustsFontSizeToFitWidth = true
-            if indexPath.row == 1 {
-                cell.textLabel?.numberOfLines = 5
-            }
-            cell.textLabel?.text = userData[indexPath.row]
-            break
-        case 1:
-            let consent = consents[indexPath.row]
-            cell.textLabel?.adjustsFontSizeToFitWidth = false
-            cell.textLabel?.text = consent
-            break
-        default:
-            break
-        }
-        return cell
     }
 }