+++ Package.swift
@@ -17,8 +17,12 @@ let package = Package(
       name: "ConsentViewController",
       dependencies: [],
       path: "ConsentViewController",
+      exclude: [
+        "Assets/javascript/GDPRJSReceiver.spec.js",
+        "Assets/javascript/jest.config.json",
+      ],
       resources: [
-        .copy("Assets/GDPRJSReceiver.js")
+        .process("Assets/javascript/GDPRJSReceiver.js")
       ]
     )
   ],