+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 5.3.2 (Nov, 23, 2020)
+* Fixed a regression that'd prevent the SDK from running on projects using Swift Package Manager #266
+
+## 5.3.2 (Nov, 23, 2020)
+* Fixed a regression in the _sharing consent with the webview_ new feature.
+
+## 5.3.1 (Nov, 23, 2020)
+* Fixed an issue that would prevent the `consentLanguage` field from the consent string to always be set to `EN`. #260
+* Added a feature to ease sharing consent between native and webviews. Check how to use it in this [section of the README](https://github.com/SourcePointUSA/ios-cmp-app#sharing-consent-with-a-wkwebview). #263
+* Fixed an issue that prevented the property `.userConsents` from the `GDPRConsentViewController` to be outdated after calling the `.customConsentTo` method. #264
+* Updated the README and the AuthExample app.
+
 ## 5.3.0 (Oct, 16, 2020)
 * Added support to Swift Package Manager (SPM).